backend/requirements.txt
fastapi
uvicorn[standard]
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pyjwt

# Database
supabase

# Queue
inngest
